CPU-Freq-Remove Header Only
